requirements.txt
numpy~=1.24.2
opencv-python~=4.6.0.66
lxml~=4.9.2
pdf2image~=1.16.3
scikit-learn~=1.2.2
setuptools~=65.5.1
layoutparser~=0.0.0
paddleocr~=2.6.1.3
matplotlib~=3.7.1
tensorflow~=2.12.0
protobuf==3.20.3
Pillow~=9.4.0
six~=1.16.0
scipy~=1.10.1
PyYAML~=6.0
tqdm~=4.65.0
pandas~=1.5.3
scikit-image~=0.20.0
h5py~=3.8.0
tensorboard~=2.12.2
keras~=2.12.0
networkx~=3.0
Flask~=2.2.3